expand validator interface
